Add CI workflows for parked app codelab
This commit adds two new CI workflows to automate the build
process for the 'Build and test a parked app for Android
Automotive OS' codelab. These workflows are triggered manually
with workflow_dispatch and on push and pull request events to the
`main` branch.

The workflows perform the following steps:

- Checkout the code
- Set up JDK 17
- Set up Gradle
- Make gradlew executable
- Build the automotive app or the regular app